build: fix the warnings the tree had, and add an opt-in -Werror
A clean -Wall -Wextra -Wconversion build over the engine, tools and tests.
Three of these were real narrowing bugs rather than noise:

  - position::do_move computed the castling mask in int and assigned back to
    a U16, so the result was silently truncated on the way in.
  - tt::save was handed qsdepth as an int where the entry field is U8.
  - uci bench computed nodes*1000.0/ms with the multiply in U64 before the
    division, losing precision on the nps figure it prints.

The rest are unused parameters, now spelled as unnamed parameters with a note
on why the argument exists at all -- the three move-scoring functions share
one signature so they can be used interchangeably as a callback, and qsearch
takes a depth it deliberately does not consult. Four memsets on
non-trivially-copyable types become value-initialisation, which GCC lowers
back to a memset.

-Wsign-conversion is deliberately left off. It fired at roughly a hundred
sites, every one a container subscripted by a non-negative engine index that
is correct by construction; a hundred warnings in a build log is how a real
one goes unread. It also does not catch what it appears to: the one genuine
out-of-bounds in this codebase (battks[65], see previous commit) is a raw
array indexed by an enum and produces no such warning. ASan found it in the
first second of the first run. Clang folds sign-conversion into -Wconversion,
so it is disabled explicitly there to keep both compilers on one policy.

HAVOC_WERROR is off by default so a local build is never blocked by a warning,
and on in CI. Verified with gcc 13.3 at -DHAVOC_NATIVE=OFF: zero warnings,
133/133 tests pass, and bench is byte-identical at 697583 nodes, so none of
this changes what the engine searches.

+++ src/uci.cpp
@@ -198,7 +198,8 @@ bool parse_command(const std::string& input, SearchEngine& engine, position& uci
             double ms = std::chrono::duration<double, std::milli>(end_time - start).count();
 
             std::cout << "Bench: " << total_nodes << " nodes, " << static_cast<int>(ms) << " ms, "
-                      << static_cast<U64>(total_nodes * 1000.0 / ms) << " nps" << std::endl;
+                      << static_cast<U64>(static_cast<double>(total_nodes) * 1000.0 / ms)
+                      << " nps" << std::endl;
         } else if (cmd == "exit" || cmd == "quit") {
             engine.stop();
             engine.wait();
